requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
gunicorn==23.0.0
python-dotenv==1.0.1
openpyxl==3.1.5
